requirements.txt
python-dotenv
groq
flask
flask-cors
